% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateVEP.R
\name{evaluateVEP}
\alias{evaluateVEP}
\title{Evaluates VEP with Statistical Measure}
\usage{
evaluateVEP(vep, stats = "pearson", label = "pearson_corr", fnc_stats = NA)
}
\arguments{
\item{vep}{A dataframe with training data to be evaluated.The first column
should be data sizes, the second and third are predicted values for
non-augmented and augmented training data. The last one should be actual
values the model is predicting.}

\item{stats}{A statistical measure to calculate. This has to be either
"pearson", "kenall" or "mse". "pearson" is set as a default.}

\item{label}{A label for the statistical measure. Used as a column name of
returned dataframe.}

\item{fnc_stats}{A function to conduct user-defined statistical analysis. The
first argument should be a dataframe to analyse, with their data size the
same. The second one is method, which does not have to be involved, just for
consistency. The thrid argument should be the data size given in the first
argument, and the last argument is label, which is the same as the parameter
above. If the parameter stats is given, fnc_stats should be NA.}
}
\value{
Returns a dataframe with three columns. The first one is the size of
training data, the second and the third columns are statistical measures for
non-augmented or augmented VEP respectively. Each row represents VEP trial.
}
\description{
A function to calculate a given statistical measure with a training data for
variant effect prediction (VEP).
}
\examples{
# Example 1:
# Evaluate VEP with pearson correlation. sample_random_data.csv is used as a
# training dataset.

df_random_data <- generateRandomData()
pearsonVEP <- evaluateVEP(df_random_data)
pearsonVEP

# Example 2:
# Evaluate VEP with mean squared error.sample_random_data.csv is used as a
# training dataset.

df_random_data <- generateRandomData()
mseVEP <- evaluateVEP(df_random_data,
                      stats = "mse",
                      label = "mean_squared_error",
                      fnc_stats = NA)
mseVEP

}
\references{
Bonnet, D.G., Wright, T.A., (2000) Sample size requirements for estimating
pearson, kendall and spearman correlations. *Psychometrika* 65, 23–28.
<https://doi.org/10.1007/BF02294183>
}
