% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomData.R
\name{generateRandomData}
\alias{generateRandomData}
\title{Generate Random Sample Dataset}
\usage{
generateRandomData(smallest = 100, largest = 1000, num_data = 1000, step = 50)
}
\arguments{
\item{smallest}{A integer that shows the smallest training data size. Default
is set 100.}

\item{largest}{A integer that shoes the largest training data size. Default
is set 1000.}

\item{num_data}{A size of the generated sample data. Default is set 1000.}

\item{step}{A step of training data size between the parameters smallest and
largest. Default is set 50.}
}
\value{
Returns a dataframe with four columns. The first column is the
training data size, second and third are predicted values for non-augmented
and augmented model. The last column is an actual value.
}
\description{
A function to generate random sample training data.
}
\examples{
# Example 1:
# Produce a sample random dataset and save as csv file.

df_random_data <- generateRandomData()
write.csv(df_random_data, "./data/sample_random_data.csv", row.names = FALSE)

}
