% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEVEP.R
\name{runEVEP}
\alias{runEVEP}
\title{Launch Shiny App for EVEP}
\usage{
runEVEP()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for EVEP
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

EVEP::runEVEP()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
