% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleRandomData}
\alias{sampleRandomData}
\title{Sample VEP file.}
\format{
A dataframe with 1000 rows.
\describe{
   \item{data_size}{A size of training data.}
   \item{non_augmented_value}{A predicted value with non-augmented training
   data}
   \item{augmented_value}{A predicted value with augmented training data}
   \item{actual_value}{A actual value the machine learning model is
   predicting}
}
}
\source{
generated by Shuzo Fukunaga
}
\usage{
sampleRandomData
}
\description{
A ramdomly generated data for VEP. Includes training data size, predicted
values for non-augmented and augmented training data model and actual value.
Contains 1000 data.
}
\examples{
\dontrun{
sampleRandomData
}
}
\keyword{datasets}
