% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEvaluation.R
\name{visualizeEvaluation}
\alias{visualizeEvaluation}
\title{Visualize VEP Evaluation}
\usage{
visualizeEvaluation(
  infile,
  stats = "pearson",
  label = "pearson correlation",
  fnc_stats = NA
)
}
\arguments{
\item{infile}{A path to a training data to be evaluated.}

\item{stats}{A statistical measure to calculate. This has to be either
"pearson", "kenall" or "mse". "pearson" is set as a default.}

\item{label}{A label for the statistical measure. Used as a column name of
returned dataframe.}

\item{fnc_stats}{A function to conduct user-defined statistical analysis. The
first argument should be a dataframe to analyse, with their data size the
same. The second one is method, which does not have to be involved, just for
consistency. The thrid argument should be the data size given in the first
argument, and the last argument is label, which is the same as the parameter
above. If the parameter stats is given, fnc_stats should be NA.}
}
\value{
Returns NA, as the graph is shown by this function.
}
\description{
A function to visualize the relationships between accuracy of VEP and
training data size, or and data augmentation.
}
\examples{
# Example 1:
# Visualize the evaluation of VEP with pearson correlation.
# sample_random_data.csv is used as a training dataset.

visualizeEvaluation("/Users/shuzo/shuzo/UofT/BCB410/EVEP/data/sample_random_data.csv")

# Example 2:
# Visualize the evaluation of VEP with mean squared error.
# sample_random_data.csv is used as a training dataset.

visualizeEvaluation("/Users/shuzo/shuzo/UofT/BCB410/EVEP/data/sample_random_data.csv",
                      stats = "mse",
                      label = "mean_squared_error",
                      fnc_stats = NA)

}
\references{
Bonnet, D.G., Wright, T.A., (2000) Sample size requirements for estimating
pearson, kendall and spearman correlations. *Psychometrika* 65, 23–28.
<https://doi.org/10.1007/BF02294183>
}
